Movimiento

+++ src/modelo/CoorAnimal.java
@@ -3,11 +3,11 @@ package modelo;
 public class CoorAnimal {
 	
 	String type;
-	int posx, posy;
+	float posx, posy;
 	
 	public CoorAnimal() {}
 	
-	public CoorAnimal(int posx, int posy, String type) {
+	public CoorAnimal(float posx, float posy, String type) {
 		super();
 		this.posx = posx;
 		this.posy = posy;
@@ -22,19 +22,19 @@ public class CoorAnimal {
 		this.type = type;
 	}
 
-	public int getPosx() {
+	public float getPosx() {
 		return posx;
 	}
 
-	public void setPosx(int posx) {
+	public void setPosx(float posx) {
 		this.posx = posx;
 	}
 
-	public int getPosy() {
+	public float getPosy() {
 		return posy;
 	}
 
-	public void setPosy(int posy) {
+	public void setPosy(float posy) {
 		this.posy = posy;
 	}
